# Fix unbounded prompt growth/determinism in scripts that loop