@dfaker 2022-11-18
Seems more like an issue with the scripts if this occurs, can you offer an example of a script causing this?