@leppie 2022-11-18
> Seems more like an issue with the scripts if this occurs, can you offer an example of a script causing this?

See `Affects` above.

I guess it could be fixed on the scripts side, but that would mean every looping script would have to have it. 
Also, not sure if the effect on `batch size` can be solved that way. 

Like said, I dont really know anything of the codebase (or Python), and I just isolated the issue. 

Not sure if that can affect other parts/scripts I have not tried.

Feel free to make alternative suggestions :D