@dfaker 2022-11-18
The style is being repeatedly appended to the prompt do we think?