@leppie 2022-11-18
> The style is being repeatedly appended to the prompt do we think?

`apply_styles` is called on ever iteration mutating the prompt, and it when it grows large, there are massive slow downs. 

See https://github.com/AUTOMATIC1111/stable-diffusion-webui/blob/eeb07330131012c0294afb79165b90270679b9c7/modules/processing.py#L454

Do a loopback script 25 times (or SD upscale 5x5) to see the effect. Make sure the promt in the style is large (> 100), 
Also do comparitive tests, using a style or the same just entered into the textbox to see that it produces different results when it should be the same.