@dfaker 2022-11-19
Grand, this seems to preserve the full substituted/appended version of the prompt everywhere it's visible too, nice catch.