@leppie 2022-11-19
> Grand, this seems to preserve the full substituted/appended version of the prompt everywhere it's visible too, nice catch.

I am still not sure if this would affect anything accidentally relying on the existing behavior. 

That said, I have been running this without issues in my normal workflow.